Move Bukkit types to lib dir
