+++ dst/lib/bukkit.js
@@ -1,0 +1,11 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var magik = magikcraft.io;
+exports.PotionEffect = magik.type("potion.PotionEffect");
+exports.PotionEffectType = magik.type("potion.PotionEffectType");
+exports.EventPriority = magik.type("event.EventPriority");
+exports.PlayerItemConsumeEvent = magik.type("event.player.PlayerItemConsumeEvent");
+exports.PlayerQuitEvent = magik.type("event.player.PlayerQuitEvent");
+exports.EventCallback = Java.type("io.magikcraft.EventCallback");
+exports.Material = Java.type("org.bukkit.Material");
+exports.PotionMeta = Java.type("org.bukkit.inventory.meta.PotionMeta");
